* change method of decrytption from tmpfile to memstream   - see also https://github.com/edp1096/php_screw1.5-bab2/   - additional fixed memory leak on zcodecom api

+++ php_screw.h
@@ -1,5 +1,5 @@
 #define PM9SCREW        "\tPM9SCREW\t"
 #define PM9SCREW_LEN     10
 
-char *zdecode(char *inbuf, int inbuf_len, int *resultbuf_len);
-char *zencode(char *inbuf, int inbuf_len, int *resultbuf_len);
+char *zdecode(char *inbuf, ULong inbuf_len, ULong *resultbuf_len);
+char *zencode(char *inbuf, ULong inbuf_len, ULong *resultbuf_len);
